Options Environment Class

The options environment stores primarily, the default values for options. In addition, it stores metadata pertaining to each option in the form of attributes.

Usage

options_initialized(env, inherits = FALSE)

init_options_env(env = parent.frame())

as_options_list(x, ...)

# S3 method for class 'list'
as_options_list(x, ...)

# S3 method for class 'options_env'
as_options_list(x, ...)

get_option_default_value(x, env = parent.frame())

get_options_spec(env = parent.frame())

get_option_spec(
  name,
  env = parent.frame(),
  inherits = FALSE,
  on_missing = warning
)

set_option_spec(name, details, env = parent.frame())

Arguments

env

An environment in which to search for an options environment

inherits

Whether to search upward through parent environments

...

Additional arguments unused

Functions

options_initialized(): Test whether options is initialized in environment

init_options_env(): Initialize an options object

as_options_list(): Convert into an options list

get_option_default_value(): Get the option's default value

get_options_spec(): Get all options specifications as named list

get_option_spec(): Get single option specification

set_option_spec(): Set single option specification

Attributes

spec: A named list of option specifications

option_name_fn: A function used to derive default option names for newly defined options. See set_option_name_fn().

envvar_name_fn: A function used to derive default environment variable names for newly defined options. See set_envvar_name_fn().